refactor: single rfl tactic
previously, `rfl` was backed by a series of `macro_rules` tactics that
tried `Eq.rfl`, `HEq.rfl`, `Iff.rfl` and an extensible tactic trying
`@refl` lemmas.

A downside of this approach is insufficient control over the error
message, and given that `rfl` is a beginner-level tactic, an improvement
here seems advisible.

This removes the `macro_rules` branching and backs `rfl` by a single
tactic that knows how to handle `Eq` directly, and then resorts to
`@refl` tactics. In all cases it does the `isDefEq` check itself (does
not leave it to `apply` to try out `@refl` lemams), so we get a
uniform error messages if the goal isn't reflexive.

Drops support for `HEq.rfl` from the `rfl` tactics; the `HEq` type's
argument order isn’t right to fit this scheme. Users are now expeced to
to use `exact HEq.rfl` manually.