feat: apply_rfl tactic: Handle Eq, HEq, better error messages
this implements the first half of #3302: It improves the extensible
`rfl` tactic (the one that looks at `refl` attributes) to

* Check itself that the lhs and rhs are defEq, and give a nice
  consistent error message when they don't (instead of just passing on
  the less helpful error message from `apply Foo.refl`)
* Also handle `Eq` and `HEq` (built in) and `Iff` (using the attribute)

Care ist taken that, as before, the transparency setting affects
comparing the lhs and rhs, but not the reduction of the relation

A test file checks the various failure modes and error messages.

I believe this `apply_rfl` can serve as the only implementation of
`rfl`, which would then complete #3302, but that seems to require a
non-trivial bootstrapping dance, so maybe better done separately.