feat: hovering over binders shows their types
Modifies `withBindingBodyUnusedName` to annotate the syntax for the variable with its corresponding fvar. Now, for example, you can hover over the variables in `fun x y => ...` in the infoview to see their types.

Also adds such annotations for the variables in `let` and `let_fun`.

Implementation note: since the fvar is local to `withBindingBody'`, this uses a fresh position for the fvar using `nextExtraPos`. This makes it safe to use `withBindingBody'` for the same expression repeatedly, since each generated fvar gets its own position.