feat: replace unary Nat.succ simp rules with simprocs
This introduces a simproc for equality and another
for subtraction so that we can remove simp attributes
on Nat.succ_sub_succ_eq_sub and Nat.succ.injEq.

Both of these have the potential to trigger stack overflows
or performance problem when applied to terms with large Nat
constants.  The simproc operations bypass this through use
of lemmas.