feat: BaseIO.asThunk
this adds
```
asThunk (f : BaseIO α) : BaseIO (Thunk α)
```
with `unsafeInterleaveIO`-like semantiks: The `IO` action will not be
executed until the first use of `Thunk.get`.

This is a samewhat unsavory operation: Whether `f` is executed now
depends on when and whether the pure `Thunk.get` is performed.
Nevertheless it seemed to be the least painful way to achieve lazy
calcuation of MessageData (see #3234), in particular because Lean's
positivity checker will allow recursion of `MessageData` through
`Thunk`, but not through `IO`.