fix: solveByElim would add symm hypotheses to local context and make impossible-to-elaborate terms
Rather than adding symm hypotheses to the local context, it now adds them to the list of hypotheses derived from the local context.

This is not idea for performance reasons, but it at least closes #3922.

In the future, solveByElim could maintain its own cache of facts that it updates whenever it does intro.