fix: mark well-founded definitions irreducible
we keep running into examples where working with well-founded recursion
is slow becuase defeq checks (which are all over the place, including
failing ones that are back-tracked) unfold well-founded definitions.

The definition of a function defined by well-founded recursion should be
an implementation detail that should only be peeked inside by the
equation generator and the functional induction generator.

We now mark the mutual recursive function as irreducible, and use
`withAtLeastTransparency .all` when producing the equations.

This fixes #3988