fix: rework elaboration interruption
Make sure interruption is never caught inside the elaborator as that
would waste CPU and in particular could expose the interrupted state to
the user if the state is reused by a later document version. Also, use a
cancellation token as the interruption source to avoid the cancellation
accidentally being inherited by request handler tasks and in preparation
for parallelism.