fix: use correct positions of patterns in `match` when delaborating
In the following, hovering over `true` in the infoview was showing `Nat.succ y`.
```lean
#check fun (x : Nat) =>
  match h : x with
  | 0 => false
  | y + 1 => true
```
Now hovering over `true` shows `true`.

The issue was that SubExpr positions were not being tracked for patterns and the position for a pattern could coincide with the pattern for a RHS. Now the position given to a pattern is an argument to a motive in a minor premise in the type of the head of the match expression.