fix: FunInd: Introduce HEq where necessary, and no equality on Props
this fixes #4146

The resulting functional induction principles generated when there is a
`match` in the function definition are far from pretty and contain
suprious copies of local definitions at their non-refined or refined
type. I expect that sometimes these may be useful, so I am hesitant to
remove them too aggressively, so for now it’s better to have too many
assumptions than too few, or even a failure (as in #4146).