fix: make erased names in `simp` clickable
as usually, just a matter of using the `WithInfo` variant.

Also simplifying the code a bit, it seems we can use
`realizeGlobalConstNoOverloadWithInfo` here.

(It's dubious API design that of all the
```
{resolve/realise}GlobalConst{NoOverload,}{WithInfo,}
```
functions the one with the longest name is the one that should be used
unless one has a reason to use another one.